refactor(app): extract prompt keydown handler to prompt-input/keydown.ts
Move handleKeyDown (the 169-line key dispatch switch) into
createPromptKeydownHandler factory. The handler routes events to popover,
history, shell-mode, file-pick, and submit/abort subsystems based on the
key combination.

PromptInput passes 17 deps to the factory, all of which were already
local closures in the handler — preserves behavior verbatim.
